# apps/backend/synapse/contracts/synapse-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Backend API",
    "version": "1.0.0",
    "description": "Shared REST contract implemented by Synapse EE and Cerebro."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/audit_logs": {
      "get": {
        "tags": ["Audit Logs"],
        "summary": "[Deprecated] List tenant audit logs — use /v1/admin/audit-log",
        "deprecated": true,
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Audit logs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditLogListResponse" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/v1/admin/audit-log": {
      "get": {
        "tags": ["Audit Logs"],
        "summary": "List tenant audit log (cursor-paginated)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 200 } },
          { "name": "before_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "principal", "in": "query", "schema": { "type": "string" } },
          { "name": "event_type", "in": "query", "schema": { "type": "string" } },
          { "name": "resource_type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Audit log page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditLogListResponse" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "responses": {
      "Forbidden": { "description": "Forbidden" },
...
    },
    "schemas": {
...
      "AuditLogListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "required": ["data"]
      },
...
    }
  }
}